refactor: small refactor of code
n
