+++ src/components/button/__tests__/button.tsx
@@ -2,7 +2,7 @@ import { describe, expect, test, vi } from "vitest";
 import { screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 
-import { render } from "@test-utils";
+import { render } from "@/utils/test-utils";
 import { Button } from "@/components";
 
 describe("Button component", () => {
