+++ src/pages/404/404.test.tsx
@@ -1,6 +1,6 @@
 import { expect, describe, test } from "vitest";
 
-import { render } from "@test-utils";
+import { render } from "@/utils/test-utils";
 import NotFound from "@/pages/404";
 
 describe("NotFound component", () => {
