+++ src/pages/about/about.test.tsx
@@ -1,6 +1,6 @@
 import { expect, describe, test } from "vitest";
 
-import { render } from "@test-utils";
+import { render } from "@/utils/test-utils";
 import About from "@/pages/about";
 
 describe("About component", () => {
