+++ tsconfig.json
@@ -24,14 +24,8 @@
     "noUncheckedSideEffectImports": true,
 
     "paths": {
-      "@/*": ["./src/*"],
-      "@test-utils": ["./utils/test-utils.tsx"]
+      "@/*": ["./src/*"]
     }
   },
-  "include": [
-    "src",
-    "vite.config.ts",
-    "eslint.config.js",
-    "utils/test-utils.tsx"
-  ]
+  "include": ["src", "vite.config.ts", "eslint.config.js"]
 }
