+++ vite.config.ts
@@ -8,7 +8,6 @@ export default defineConfig({
   resolve: {
     alias: {
       "@": path.resolve(__dirname, "./src"),
-      "@test-utils": path.resolve(__dirname, "./utils/test-utils.tsx"),
     },
   },
   server: {
